chore: Update public speaking event title
Changed the title of the public speaking event from "Ancillary Services and the Green Energy Transition" to "A future perspective on Ancillary Services and the Green Energy Transition" for clearer communication of the topic. This update was made to better reflect the content of the presentation and capture the audience's attention.